+++ ApplicationTracker.Services/Interfaces/IService.cs
