+++ ApplicationTracker.Services/JobTitleService.cs
@@ -2,50 +2,49 @@
 using ApplicationTracker.Data.Dtos;
 using ApplicationTracker.Data.Entities;
 using ApplicationTracker.Services.Interfaces;
-using System.Runtime.CompilerServices;
-using Serilog;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 
 namespace ApplicationTracker.Services
 {
-    public class WorkEnvironmentService(TrackerDbContext context, ILogger<WorkEnvironmentService> logger) : IService<WorkEnvironmentDto>
+    public class JobTitleService(TrackerDbContext context, ILogger<JobTitleService> logger) : IService<JobTitleDto>
     {
         private readonly TrackerDbContext _context = context;
-        private readonly ILogger<WorkEnvironmentService> _logger = logger;
+        private readonly ILogger<JobTitleService> _logger = logger;
 
-        public async Task<IEnumerable<WorkEnvironmentDto>> GetAllAsync()
+        public async Task<IEnumerable<JobTitleDto>> GetAllAsync()
         {
-            return await _context.WorkEnvironments
-                .Select(x => new WorkEnvironmentDto { Id = x.Id , Name = x.Name })
+            return await _context.JobTitles
+                .Select(x => new JobTitleDto { Id = x.Id, Name = x.Name })
                 .ToListAsync();
         }
 
-        public async Task<WorkEnvironmentDto?> GetByIdAsync(int id)
+        public async Task<JobTitleDto?> GetByIdAsync(int id)
         {
-            var result = await _context.WorkEnvironments
+            var result = await _context.JobTitles
                 .Where(x => x.Id == id)
-                .Select(x => new WorkEnvironmentDto { Id = x.Id, Name = x.Name })
+                .Select(x => new JobTitleDto { Id = x.Id, Name = x.Name })
                 .FirstOrDefaultAsync();
 
             if (result == null)
             {
-                _logger.LogInformation("No WorkEnvironment with id {id} found", id);
+                _logger.LogInformation("No JobTitle with id {id} found", id);
             }
             return result;
         }
 
         public async Task<bool> ExistsAsync(int id)
         {
-            return await _context.WorkEnvironments.AnyAsync(x => x.Id == id);
+            return await _context.JobTitles.AnyAsync(x => x.Id == id);
         }
 
         public async Task<IEnumerable<ApplicationDto>?> GetRelatedApplicationsAsync(int id)
         {
             return await _context.Applications
-                .Where(x => x.WorkEnvironmentId == id)
-                .Select(x => new ApplicationDto
-                {
-                    ApplicaitionDate = x.ApplicationDate,
+                .Where(x => x.JobTitleId == id)
+                .Select(x => new ApplicationDto 
+                { 
+                    ApplicaitionDate = x.ApplicationDate, 
                     Source = new SourceDto { Id = x.SourceId, Name = x.Source.Name },
                     Organization = new OrganizationDto { Id = x.OrganizationId, Name = x.Organization.Name },
                     JobTitle = new JobTitleDto { Id = x.JobTitleId, Name = x.JobTitle.Name },
