+++ ApplicationTracker.Services/LocationService.cs
@@ -2,44 +2,46 @@
 using ApplicationTracker.Data.Dtos;
 using ApplicationTracker.Services.Interfaces;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 
 namespace ApplicationTracker.Services
 {
-    public class OrganizationService(TrackerDbContext context, ILogger<OrganizationService> logger) : IService<OrganizationDto>
+    public class LocationService(TrackerDbContext context, ILogger<LocationService> logger) : IService<LocationDto>
     {
         private readonly TrackerDbContext _context = context;
-        private readonly ILogger<OrganizationService> _logger = logger;
+        private readonly ILogger<LocationService> _logger = logger;
 
-        public async Task<IEnumerable<OrganizationDto>> GetAllAsync()
+        public async Task<IEnumerable<LocationDto>> GetAllAsync()
         {
-            return await _context.Organizations
-                .Select(x => new OrganizationDto { Id = x.Id, Name = x.Name })
+            return await _context.Locations
+                .Select(x => new LocationDto { Id = x.Id, Name = x.Name })
                 .ToListAsync();
         }
 
-        public async Task<OrganizationDto?> GetByIdAsync(int id)
+        public async Task<LocationDto?> GetByIdAsync(int id)
         {
-            var result = await _context.Organizations
+            var result = await _context.Locations
                 .Where(x => x.Id == id)
-                .Select(x => new OrganizationDto { Id = x.Id, Name = x.Name})
+                .Select(x => new LocationDto 
+                    { Id = x.Id, Name = x.Name, City = x.City, State = x.State, Country = x.Country })
                 .FirstOrDefaultAsync();
 
             if(result == null)
             {
-                _logger.LogInformation("No Organizations with id {id} found", id);
+                _logger.LogInformation("No Location with id {id} found", id);
             }
             return result;
         }
 
         public async Task<bool> ExistsAsync(int id)
         {
-            return await _context.Organizations.AnyAsync(x => x.Id == id);
+            return await _context.Locations.AnyAsync(x => x.Id == id);
         }
 
         public async Task<IEnumerable<ApplicationDto>?> GetRelatedApplicationsAsync(int id)
         {
             return await _context.Applications
-                .Where(x => x.OrganizationId == id)
+                .Where(x => x.LocationId == id)
                 .Select(x => new ApplicationDto
                 {
                     ApplicaitionDate = x.ApplicationDate,
